CORE-8643 Move Properties files to model package
