+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/model/FileInfoTypeProperties.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.apps.integration.client.view.propertyEditors;
+package org.iplantc.de.apps.integration.client.model;
 
 import org.iplantc.de.client.models.apps.integration.FileInfoType;
